requirements-domain-eval.txt
# Additional dependencies for domain evaluation
sentence-transformers
faiss-gpu
bert-score
scikit-learn
pandas